action and billing now appear to be working
