+++ billing.py
@@ -20,6 +20,7 @@ import click
 # TEST database call
 # from sqlalchemy import create_engine, MetaData, String, Table, Float, Column, select
 import logging
+from database import db, login_db, login_manager, User
 
 billingbp = Blueprint('billingbp', 'billingbp')
 
@@ -28,9 +29,16 @@ logging.basicConfig()
 es_service = os.environ.get("ES_SERVICE", "localhost")
 es = Elasticsearch(['http://' + es_service + ':9200/'])
 
-#@login_manager.user_loader
-#def load_user(user_id):
-#    return User.query.get(int(user_id))
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
+
+@billingbp.route('/login')
+def login():
+    if current_user.is_authenticated:
+        redirect('https://{}'.format(os.getenv('DCC_DASHBOARD_HOST')))
+    else:
+        redirect('https://{}/login'.format(os.getenv('DCC_DASHBOARD_HOST')))
 
 @billingbp.route('/invoices')
 @login_required
