+++ mapi.py
@@ -57,30 +57,29 @@ class Config(object):
 
 """ DB Models """
 
-class User(login_db.Model, UserMixin):
-    __tablename__ = "users"
-    __bind_key__ = "login-db"
-    id = login_db.Column(login_db.Integer, primary_key=True)
-    email = login_db.Column(login_db.String(100), unique=True, nullable=False)
-    name = login_db.Column(login_db.String(100), nullable=True)
-    avatar = login_db.Column(login_db.String(200))
-    access_token = login_db.Column(login_db.String(5000))
-    redwood_token = login_db.Column(login_db.String(5000))
-    tokens = login_db.Column(login_db.Text)
-    created_at = login_db.Column(login_db.DateTime, default=datetime.datetime.utcnow())
+#class User(login_db.Model, UserMixin):
+#    __tablename__ = "users"
+#    __bind_key__ = "login-db"
+#    id = login_db.Column(login_db.Integer, primary_key=True)
+#    email = login_db.Column(login_db.String(100), unique=True, nullable=False)
+#    name = login_db.Column(login_db.String(100), nullable=True)
+#    avatar = login_db.Column(login_db.String(200))
+#    access_token = login_db.Column(login_db.String(5000))
+#    redwood_token = login_db.Column(login_db.String(5000))
+#    tokens = login_db.Column(login_db.Text)
+#    created_at = login_db.Column(login_db.DateTime, default=datetime.datetime.utcnow())
 
 app = Flask(__name__)
 app.config['DEBUG'] = True
 app.config.from_object(Config)
-app.register_blueprint(actionbp)
-app.register_blueprint(webservicebp)
-app.register_blueprint(billingbp)
 db.init_app(app)
 login_db.init_app(app)
 login_manager.init_app(app)
+app.register_blueprint(actionbp)
+app.register_blueprint(webservicebp)
+app.register_blueprint(billingbp)
 apache_path = os.environ.get("APACHE_PATH", "")
 es_service = os.environ.get("ES_SERVICE", "localhost")
-login_manager = LoginManager(app)
 login_manager.login_view = "login"
 login_manager.session_protection = "strong"
 migrate = Migrate(app, db)
@@ -90,5 +89,3 @@ app.app_context()
 
 if __name__ == '__main__':
     app.run()  # Quit the debu and added Threaded
-
-
